How Do You Take Care Of Body Wave Sew In?
Body wave hair, one of the most popular hair textures in wigs and hair bundles, caters to a wide range of women's styles and looks.
Though body wave sew-in hair has a shine, it can become dull and lose its waves if not adequately cared for. So, once you've received your sew-ins, the care routine begins. Human hair and synthetic hair are the two main types of body wave sew-in. Human hair has a more natural appearance and lasts longer than synthetic hair, but it is more expensive. As a result, it requires time and energy to maintain the investment and keep the hair alive and shiny.
Before we begin discussing how to care for body wave hair, we must first define what body wave hair is.